Don't keep checking if migration is needed
This has serious performance penalties when the imap file is large as the check is being run for each message imported. With a 20MB imap file (approximately 300,000) emails, it takes about 0.5 seconds to check the migration status on each message import.

Instead, we remember that we have successfully validated and skip the check on subsequent calls to `validate!`. If validation fails, we allow the method to evaluate fully on subsequent calls.