Ensure version 2->3 migration happens before any .imap access
The delegators were allowing access to the metadata before it
was migrated from the legacy version.
This meant that the list of UIDs was empty.
The second problem was that the Serializer::Imap instance
was not reloaded. This meant that it continued not be '#valid?'.
For this reason, after version migration both the '.imap' and '.mbox'
files were deleted.

Now, the migration happens on the first attempt at access to the
metadata and the fresh metadata is then loaded.