Add 'env' parameter for the 'backup' command
This commit addresses
[#issue 167](https://github.com/joeyates/imap-backup/issues/167)

Passing `--env` to the `backup` commands enables the use of environment
variables for passwords in the configuration file.

It allows users to not store their password in plain text in the
configuration file.

Passwords in the configuration file which are valid environment variable
names preceeded by a dollar sign are replaced by the value of the
corresponding environment variable.
For example: `"password" : "$PASSWORD"`

When the environment variable is not set, the user gets a password
prompt.

When the password is not a valid environment variable name preceeded by
a dollar sign, the configuration is returned as is.